Chore: widget split
- home button, bottom menu 등 widget 분리
- 일부 파일 주석 작성
